extract history model and history renderer, trim long inputs in renderer
